Update Functions Docs to Include Functions Domain
I was unaware the functions were actually being served through the browser, I thought you had to interact with them through the command line.  While trying to find a way to interact with them through the browser I read the netlify-lambda docs and realized that netlify dev does serve the files in the browser, it's just not mentioned anywhere.