Fix: Return .html file when file and folder share name
When you have a folder and a file with the same name (e.g., foo.html and
foo/), Netlify will serve the contents of foo.html. Until now,
netlify dev was inconsistent with that behavior and returned a 403
Forbidden error (because directory listings are forbidden) in this
scenario. This commit makes netlify dev behave consistent with Netlify
by serving the .html (or .htm) file instead of returning 403 Forbidden.

Fixes: https://github.com/netlify/cli/issues/1017