Fix: Use 404.html page in publish folder
In Netlify, you can set a custom 404 page by putting a 404.html file
into the publish folder (e.g., public or whatever is specified in
netlify.toml). Until now, this custom 404 page was not picked up by
netlify dev. For it to work in netlify dev, the custom 404.html file had
to be in the root folder. This commit makes the behavior of netlify dev
consistent with Netlify in production by using the custom 404.html page
in the publish folder.

Fixes: https://github.com/netlify/cli/issues/1158