Update detect-server.js
This was a very confusing error message. "args" in this case is an implementation detail, and the error message didn't provide any insight as to what might have gone wrong.

In my case, it was using `craco` to override my base `create-react-app` config, which meant a change to the start script from `react-scripts-start` to `craco start`. I didn't know how to use a custom framework until after writing a craco-specific detector in my local install of the CLI