chore(docs): update getting-started.md (#4287)
This PR updates the "locally" world to "globally", because the subsequent command installation has this global `-g` flag:

```bash
npm install netlify-cli -g
```

At least, that's how I see it. It was a bit confusing at me first when the caption is "locally" but the code snippet after has this `-g` flag. If this PR doesn't make sense, please feel free to close this. Thanks!