chore: split package size workflow to make it work for PRs from forks
This enables the package-size to work also from forks. Currently
it does not work because the workflow gets only a readonly token
if started from a fork, hence the workflow cannot post the comment.

By splitting we run the calculation in a secure environment with
only the readonly token and run the second workflow after the first
one is finished. The second workflow then has write access and can
post the comment

taken from https://securitylab.github.com/research/github-actions-preventing-pwn-requests/