fix(live): pass URL to build env instead of setting LIVE_URL
turns out that it's only builds and build plugins that have this problem
and that functions and edge functions already pick up the right URL and DEPLOY_URL
in the environment. This fixes the URL for builds and build plugins when --live is
set, but the problem still remains when not running in dev mode.