chore: fix handle-questions helper to support autocomplete (#5629)
The problem was that the inquirer autocomplete was not catching up with the date
we wrote to stdin. By giving it 50ms breathing room before we press ENTER the autocomplete
does work.

This also chooses the set-cookie template for the functions-create test
as it only has the `cookie` dependency to install instead of the whole apollo server previously.

Hopefully this will not make these tests timeout so much.

Co-authored-by: kodiakhq[bot] <49736102+kodiakhq[bot]@users.noreply.github.com>