fix(recipes): suggest a replacement on typo again
When the user specifies a recipe that doesn't exist, the recipes command
is supposed to suggest a replacement. The way it detects this scenario
is to catch the module not found error, but it was catching
MODULE_NOT_FOUND, which is only thrown for CommonJS modules:

> MODULE_NOT_FOUND
>
> A module file could not be resolved by the CommonJS modules loader
> while attempting a require() operation or when loading the program entry
> point.
>
> --- https://nodejs.org/api/errors.html#module_not_found

For ESM modules, when a module isn't found, the ERR_MODULE_NOT_FOUND
error is thrown instead:

> ERR_MODULE_NOT_FOUND
>
> A module file could not be resolved by the ECMAScript modules loader
> while attempting an import operation or when loading the program entry
> point.
>
> --- https://nodejs.org/api/errors.html#err_module_not_found

This commit makes it catch ERR_MODULE_NOT_FOUND instead.