feat!: support monorepo development
BREAKING CHANGE: In monorepos it is now prompting you
to select the site you want to run the operations in.

This can be bypassed by providing a `--filter` argument
and specify a site/workspace package that should be used.

Through that config file paths are changed and toml files
are picked up accordingly out of the package folders.