chore: replace ava with vitest (#6103)
* test: remove obsolete snapshots

* test: replace ava with vitest

* chore: remove ava package from dependencies

* refactor: rename files

* chore: add ava config back in

Even though we're not using ava in this package,
our internal lint package (@netlify/eslint-config-node)
doesn't recognize this.

As a result it's incorrectly failing lint checks because files
that are not test files as being imported into actual test files
(the specific rule that's failing is `ava/no-import-test-files`.

* chore: remove legacy action since it's no longer needed

update vitest command

* test: add tools into the unit tests workflow to make sure they're still run

* refactor: have to add this back in since the ava property had to be added back in

* test: forgot to add folder to vitest config

* test: remove tools from test run

we need to come back and fix these

* refactor: forgot to remove a command

* test: remove tools from vitest config

---------

Co-authored-by: Sarah Etter <sarah.etter@netlify.com>
Co-authored-by: kodiakhq[bot] <49736102+kodiakhq[bot]@users.noreply.github.com>