fix: `sites:create-template` command erroneously creating GitHub repos
Previously, if you had a site on Netlify without a corresponding GitHub repo,
running `sites:create-template` would prompt for a new name *and* create the
repo on GitHub. This fix uses a `listSites` API call to check if the site already
exists. If it does, it prompts for a new name and doesn't create a new repo.

Co-authored-by: Ben Hancock <benhancock859@gmail.com>