fix: prevent redundant repos, add preliminary name checks, and improve error handling in sites:create-template (#6908)
* fix: `sites:create-template` command erroneously creating GitHub repos

Previously, if you had a site on Netlify without a corresponding GitHub repo,
running `sites:create-template` would prompt for a new name *and* create the
repo on GitHub. This fix uses a `listSites` API call to check if the site already
exists. If it does, it prompts for a new name and doesn't create a new repo.

Co-authored-by: Ben Hancock <benhancock859@gmail.com>

* fix: re-factored and re-organized

Co-authored-by: Ben Hancock <benhancock859@gmail.com>

* fix: fixed bug in sites:create-template

Previously, if you had a site on Netlify without a corresponding
repo, `sites:create-template` would create a repo *and* prompt for
a new name, resulting in an unused repo. The same situation would
occur if the site existed globally. This fix implements checks that
avoid creating repos unnecessarily.

Co-authored-by: Ben Hancock <benhancock859@gmail.com>

* test: started test file for sites:create-template

Co-authored-by: Ben Hancock <benhancock859@gmail.com>

* fix: update sites-create-template

prevent redundant repos, add preliminary name checks,
and improve error handling in sites:create-template

Co-authored-by: Ben Hancock <benhancock859@gmail.com>

* fix: remove unneccessary comments

Co-authored-by: Ben Hancock <benhancock859@gmail.com>

* fix: increase timeouts to fix tests

Co-authored-by: Ben Hancock <benhancock859@gmail.com>

* fix: solves test bugs

Co-authored-by: Ben Hancock <benhancock859@gmail.com>

* fix: rewrote tests

Co-authored-by: Ben Hancock <benhancock859@gmail.com>

* fix: remove unneccessary mocks

Co-authored-by: Ben Hancock <benhancock859@gmail.com>

---------

Co-authored-by: Ben Hancock <benhancock859@gmail.com>