feat!: remove deprecated camelCase option variants (#7046)
* feat!: remove deprecated camelCase options

These commands option have been deprecated since 12.8.0 (25 months ago):
https://github.com/netlify/cli/releases/tag/v12.8.0.

Removes:
- `--httpProxyCertificateFilename`
- `--prodIfUnlocked`
- `--functionsPort`
- `--targetPort`
- `--edgeInspect`
- `--edgeInspectBrk`
- `--replaceExisting`
- `--gitRemoteName`

The change in 12.8.0 also intended to rename `--httpProxy` to `--http-proxy` but there was a typo resulting in that
never happening. To err on the safe side, this fixes that change without removing the camel case variant yet.

* fix: update references to renamed options

* feat!: rename `--httpProxy` to `--http-proxy`

This was not a documented option, so this is arguably not breaking, but marking as
breaking to err on the side of caution.