fix!: upgrade to commander@12 and fix some conflicting short option names (#7008)
* fix: use a valid short flag for blobs:get output

`-o` is already used in the base command for `--offline`, so this was being quietly ignored

This can't be considered a breaking change as it has never worked. (I just tried it to be sure.)

See https://github.com/tj/commander.js/pull/2055.

* fix: avoid redefining `--json` option in commands

See https://github.com/tj/commander.js/pull/2055.

* fix: fix duplicate `-o` option for deploy command

It's already used in the BaseCommand for `--offline`.

This is not a breaking change, because this never worked as is.

Same issue as in 9280c558.

* fix: fix duplicate `-a` option in deploy command

It's already defined in the BaseCommand for `--offline`.

* fix: fix still more `--json` option overrides

* fix: fix env:import options

`-r` was already used globally so this was being ignored

* fix: don't mutate the commander singleton

It was redefining the options over and over and over. The latest commander throws.

* chore(deps): bump astro in docs site

* fix(deps): upgrade to commander@12

* types(site): fix jsdoc types in docs script

* fix: remove `--json`, `--offline` from BaseCommand

and centralize `--auth` only in `BaseCommand`

commander@12 doesn't allow redefining the same option, so this pattern was no longer viable

This was a pure refactor, but it uncovered some commands that weren't declaring options they were using, so I fixed
those too.

* fix: undo unnecessary short flag change