chore: simplify and standardize `package.json` scripts and CI setup (#7069)
* chore(package.json): simplify site build

There were two separate scripts for doing the same thing.

* refactor(scripts): split up prepare-for-publish

No behavioural changes in this commit.

* chore: remove bizarre postinstall hook that builds

- It's nonstandard
- It's surprising
- Since we use `tsc` to build, it causes every CI check to fail if you have type errors
- It causes issues when rewriting git history
- It's a bit magical

* chore: remove leftover refs to ava

ava was removed a couple years ago

* ci: rename misleading jobs

They handle two kinds of benchmarks, only one of which is package size

* chore: split formating, linting, and unit testing

* ci: simplify workflow skipping for release branches

* chore(package.json): sort scripts

* chore(deps): add missing dependency on npm-run-all

We're using `run-s` which comes from this dep but it only happens to work because one of
our dependencies has a transitive dependency on `npm-run-all`.

IMO we should kill this, but for now this just adds the missing dep.

Note: npm-run-all2 is a maintained fork of the unmaintained npm-run-all