refactor: modernize ESLint configuration
This changeset migrates the project to ESLint 9 and to its new
configuration format (`eslint.config.js`).

To do this, I more or less ditched our current shared configuration in
favor of using `recommended` presets from the plugins we use. We haven't
been maintaining our ESLint preset since early 2022 and it uses long-
outdated lint plugins (e.g. a version of the TypeScript ESLint plugin
that was published in mid 2023). Even though this change will likely
get rid of some rules we might find helpful and re-enable later, I think
on the whole this is a win and a foundation for making ESLint a more
useful tool going forward.

A few other noteworthy changes here:

- Cuts ESLint runtime from ~1 min to ~7s
- This update makes a few extraneous dev dependencies explicit. A lot of
  our ESLint dependencies (eslint, plugins, prettier, etc.) were
  transitive dependencies from `@netlify/eslint-config-node`, which is a
  pattern notorious for causing package resolution issues. This change
  makes these dependencies explicit.
- Replaced the lone custom ESLint rule (restricts the use of
  `process.cwd()`) with a built-in equivalent.
- Updated the `tsconfig.json` configuration to use a more accurate
  configuration for our version of node. (We specified targets, `lib`s,
  etc. that were outdated for Node 18.) This change was prompted by some
  lint errors that made no sense.
- Added the Vitest ESLint plugin.