build: don't build project in postinstall script
The CLI currently runs `npm run build` in a post-install script. Many
of commands in the CLI implicitly depend on this behavior, which is
unnecessarily complicated. This post-install script conflicts with the
user-facing post-install script, which means we need to do a little pre-
publish post-install script juggling to ensure that the end user runs
the correct script after installing the CLI.

This change removes the post-install script in favor of explicitly
building the CLI before e.g. CI steps that require a built CLI.

Partial un-reversion of https://github.com/netlify/cli/pull/7069, which
was reverted in https://github.com/netlify/cli/pull/7085.