test: don't reset workspace when running E2E tests
The E2E test script unexpectedly runs a `git checkout .` on the CLI
repository. If you run it locally and have any uncommitted changes,
oops! It just discarded your work.

https://github.com/netlify/cli/pull/7119 removes the
`package.json#scripts` pre-publish mangling that we used to do, which
is ostensibly why we needed this in the first place.