test: remove certs script
We currently have a `certs` npm script that generates self-signed
certificates for use in a single zisi test. This script must be run
before running tests, otherwise the zisi test in question will fail.

This changeset moves this certificate generation into a test helper
function, ensuring that a certificate will always be available when
needed by a test.

It would probably be simpler to instead generate a certificate that
never expires and inline it into the fixture (eliminating the need for
an available `openssl` binary), but this works fine.