ci: run all workflows when PRing to non-main branches
Most of our GitHub Actions workflows are configured to only run on PRs
to `main`.

This sucks when you're basing changes on top of each other. Consider the
following example:

```
main <-- feat/my-feature <- feat/my-sub-feature
```

As configured now, CI will never run on `feat/my-sub-feature`.

It's useful to be able to break changes up into smaller PRs for review,
or to be able to sub-branch to avoid merge conflicts while you're
waiting for a review.