ci: simplify release-please conditionals
This changeset modifies the CI test workflows to not run on release PRs.

One effect this PR will have is that these workflows will no longer
report a status on release-please PRs. Because we have our repo rules
configured to require these workflows' jobs to block merges to main
unless these workflows are in a passing state, merging this as is will
break release-please branches. To fix this, we'll modify the
merge-to-main ruleset to exclude the release-please bot user from the
ruleset's requirements.