style: lint all files
I modernized the eslint configuration Back in 25fcdd8. At the time, I
chose not to make any changes to the list of filepaths we historically
passed to eslint, thinking that maybe this would include a bunch of
files (like fixtures) that we intentionally ignore.

This turned out to be not the best idea. I didn't notice at the time
that the list of filepaths was not great to begin with: We were passing
a bunch of file extensions (mjs, cjs) that we no longer have in the
repository, and other extensions (md, html) that ESLint doesn't lint.

Since ESLint 9, you no longer ever really have reason to pass a path or
paths to `eslint` (e.g. you often used to pass `.`, but this is now the
implied default). So, this change just changes our use of ESLint to
match what the project recommends.

This change exposed a lot of unlinted files, so I've fixed or suppressed
a minimum of errors in this PR. We can fix these suppressions as we go.