ci: don't retry integration, unit test suites 3x in CI
This updates the unit test and integration test CI workflows to no
longer retry the entire test suite when one or more test suites fails.
This is in place to compensate for flaky test suites, but this approach
makes individual test suites take up to 15 minutes when a test is
legitimately failing and masks which tests are actually flaky.

I've added inline retry configurations to known flaky tests and slapped
a `--retry=3` on the integration test suite to cover the rest of the
flaky integration tests (many of them are flaky due to port conflicts).
As time goes on, hopefully we can fix and/or inline-annotate the
remaining test flakes and get rid of the global retry config on
integration tests.

In the interim, this may cause an uptick in CI flakes, but I think after
a week or two we'll be able to root out test flakes and make CI quicker
and more reliable.

Once I did this, the `get-global-config` unit tests started reliably
failing on Windows. I looked into it and discovered that the first unit
test run has always been failing on Windows. Adding test-specific
retries did not fix it; I assume that the `get-global-config` test
suite's setup is inadequate and indirectly relies on another test
correctly setting up the filesystem. I fixed this by converting the
`get-global-config` tests to use a mocked in-memory filesystem. When I
couldn't get `fs` mocks to work reliably, I looked into `configstore`
and discovered it's extremely simple and has a few dependencies we could
get rid of, so I just rewrote it.