perf!: perform minimal accounts query when loading config
We've done this performance improvement many times elsewhere, but for
posterity:

When we load a user's configuration, we perform a query to the Netlify
API requesting information about which accounts a user is a member of.
To do this, we query the `/accounts` endpoint, which is an expensive
operation: It frequently takes 4+ seconds to get a response from this
endpoint. This query happens as part of every CLI startup cycle.

By requesting a more minimal set of `accounts` data, we can shave 2-3
seconds off almost every command's startup time.