fix(types): fix 388 type errors and fully type `src/lib/functions/` (#7130)
* ci: unfortunately build before typecheck/lint

* refactor: port go/rust unit tests to TS

* refactor: fail a complicated test more gracefully

* fix(types): fix functions types and much more

* refactor: split `error()` helper to aid typing

By separating the version that throws from the version that does not, we can annotate it
the former as returning `never`, which trickles down to improving a bunch of other types
at call sites. I also cleaned up a bunch of otherwise necessary weird type workarounds
like throwing or returning some garbage after having already thrown.

I also renamed the method(s) to be clearer.

* build: remove resolved temporary suppressions

* fix: fix functions regression and fully type `lib/functions/`

The args passed to to `.postMessage()` was the regression.

* refactor: port functions-serve test file to TS

* fix: invoke functions "directly" by default

I'm not sure why/whether this is important, but it fixes this test.

Regression introduced a few commits earlier in this same PR.

This also adds a test that helped me narrow in.

* refactor: remove stray comment

* fix: avoid potential change in function behaviour

* test: test source code in .ts files

* Revert "ci: unfortunately build before typecheck/lint"

This reverts commit 391c8415cbbf56b9e8a5156c595049b1eefe1cdf.

* fix(deps): fix local functions for node <18.20.0

* fix(deps): bump local-functions-proxy for bug fix

* fix: fix typo regression in error message

* refactor: rename `getPackageJson` to clarify intent

* fix: don't traverse fs looking for package.json

I technically changed these behaviours when I pulled in `read-package-up` in these code
paths. This switches to `read-pkg` to keep the previous behaviour.

* build: fix eslint regressions from rebase

* docs: clarify ambiguous todo comments

* refactor: avoid disabling require-await eslint rule

* refactor: simplify caught error conditional

* fix: make behavioural change slightly safer

* fix: revert removal of `plugins[].inputs`

* style: fix inline code formatting hint

* fix: wrap unknown headers error at runtime

* style: consistently use colon after todo comments

* refactor: avoid a @ts-expect-error due to is-stream

* refactor: use normalize-package-data directly

* style: use nullish coalescing