perf: enable V8 code cache when available
See https://nodejs.org/api/module.html#module-compile-cache.

In local benchmarking on various machines, this led to a 75-400 ms improvement to command
run time across the board, at the expense of a moderate increase in first-time run time.

See inline - to account for the common CI single-run use case, this is disabled when we
detect a CI environment.