perf: speed up all commands by 75+ ms by enabling V8 code cache when available (`node>=22.8.0`) (#7173)
* build(deps): use correct version of @types/node

Our minimum supported version is 18, so we need to rely on node 18 types.

* refactor: wrap main program in a main function

* chore(devDeps): remove unused is-ci dev dep

* refactor: allow calling `setAnalyticsPayload` earlier

The `preAction` hook was overwriting the whole payload, so any prior calls to
`setAnalyticsPayload` would have their data thrown away.

This didn't matter, but I'd like to introduce an early call.

* perf: enable V8 code cache when available

See https://nodejs.org/api/module.html#module-compile-cache.

In local benchmarking on various machines, this led to a 75-400 ms improvement to command
run time across the board, at the expense of a moderate increase in first-time run time.

See inline - to account for the common CI single-run use case, this is disabled when we
detect a CI environment.

* fix: log telemetry track errors in debug mode

* refactor: extract compile cache module

This also removes the child process handling.

* build: disable compile cache when running tests

* perf: defer more imports in bin/run

* docs: add another clarification to compile cache