refactor: convert remaining JavaScript test files to TypeScript
This changeset converts all remaining JavaScript tests to TypeScript.
It's sort of a hack job; I used a bulk converter for most of it, then
hand-converted the rest. There's varying degrees of "hey, he did a nice
job in this file" and "hey, he did a terrible job in this file," but on
the whole this is an improvement.

I also updated all of the import paths in tests from importing
transpiled (`dist/`) code to untranspiled (`src/`) code. Ergonomically
this is better--it's confusing to work on tests that aren't necessarily
running code that reflects your edits--but really, I was forced into
this by spooky conflicts between transpiled types and untranspiled
types.