test: fix flakey snapshot normalization logic
This logic was normalizing integration test snapshots by replacing any sequence of five
digits with `88888`. The intent is to handle some randomness in test output.

It wasn't working some percentage of the time because randomly selected ports are
sometimes _four_ digits. You can see an example in this failed run:
https://github.com/netlify/cli/actions/runs/14434953424/job/40474516002?pr=7202#step:10:1143.

While I was at it, I made it less much less generic and moved it into the only test suite
that relies on it.