fix: match the `branch-deploy` context when using `--context branch:foo` (#7235)
* fix: match `branch-deploy` context when using `--context branch:foo`

The `--context` parameter can be passed to `dev`, `serve`, `functions:serve`, and
`deploy`.

As documented, its value can be a valid context (`all', 'production', 'deploy-preview',
'branch-deploy', 'dev`) or `branch:foo` where `foo` is the name of a branch.

In the former case, we match env vars with a matching `context`.

In the latter case, the intent is to match env vars with `context: "branch"` and
a `context_parameter` matching the branch name (`foo`).

All these behaviours are working as intended.

However, in addition to these, it's expected that using `--context branch:foo` implicitly
inherits the `branch-deploy` context as well (although a matching `branch` context value
should take precedence). This was not implemented.

You might think this is unnecessary: if you don't need the override, just use `--context
branch-deploy`. However, consider multiple environment variables, some of which have
branch-specific overrides, some of which do not. There is currently no mechanism available
to make this work as expected.

This commit fixes that, by simply adding the missing logic.

* fix(docs): improve inline help and docs for `--context`

* fix: use smart quotes in build --context docs docs

* fix: fix typo in deploy --context docs