fix: fix framework server loading spinner ("Waiting for framework port") (#7242)
* fix: fix framework server loading spinner rendering

This is a regression in the loading spinner shown before "Waiting for framework port", introduced in 19.0.3.

* fix: fix framework server spinner

See:
1. https://github.com/netlify/cli/pull/7026
2. https://github.com/netlify/cli/pull/7131
3. https://github.com/netlify/cli/issues/7124

Oops, I was really overthinking this. All this needs to do is, every time a chunk comes in
from the underlying framework dev server, clear the spinner line, write the chunk, then
resume the spinner.

This wasn't set up right at all, which was obfuscated by the erroneous use of `isSpinning`
(see previous commit).