feat(telemetry): send command opts and args
It's impossible for us currently to identify the potential impact of option deprecation.
This won't address that retroactively for past CLI versions, but at least it will give us
data going forward.

Attaching the `source` is important, otherwise this would include applied default values
and we wouldn't be able to distinguish this from actual user-provided values.

Unfortunately there doesn't appear to be a commander API to get named positional
arguments, so `args` is just an array. Seems ok.