feat(init): prompt to configure build settings in 'no git remote' case
We currently do this in the 'happy path' case where a git remote is found (regardless of
whether the use chooses to *connect* their site to this git remote (i.e. configure Netlify
CD), but not in the case where the user chooses the "Yes, create and deploy site manually"
answer to the "Do you want to create a Netlify site without a git repository?".

It turns out this is a pretty common choice, and it's confusing that it doesn't detect and
recommend build settings for your project and prompt you to write them to a `netlify.toml`

This ports over the same flow to this scenario.