feat: add new `netlify clone` command, to clone an existing repo + connected Netlify site in one step (#7260)
* refactor(link): extract function for site by repo URL

This also polishes the progress output and fixes some types.

* feat: add `clone` command

When a user already has an existing Netlify site connected to a hosted git repo, pulling
this down and preparing it for use with Netlify CLI currently requires numerous steps:
1. find and copy the git remote URL
2. run `git clone <git url>`
3. run `cd <created dir>`
4. run `netlify link`
5. follow prompt to select "Use current git remote origin"

This is a very common flow, e.g. when using a "Deploy to Netlify" button, creating a site
from a Netlify template, creating a site from an AI creator... any workflow where the site
is created outside Netlify CLI requires several steps to get it ready for use with Netlify CLI.

This introduces a new command which vastly simplifies these steps:
1. find and copy the git remote URL
2. run `netlify clone <git url>`

This will clone the git repo, find the corresponding Netlify site, and link the created
directory automatically. It will only prompt the user if necessary - for example if the
git repo URL matches multiple sites.

For convenience and future integration into the Netlify platform and docs, I also included
support for explicitly passing the known site id or name via `--id` or `--name`.