fix: improve link fallback when OSC 8 not supported
The `terminal-link` package uses [OSC 8](https://github.com/Alhadis/OSC8-Adoption/) link
formatting when running in a terminal that supports it. When it isn't supported, by
default it falls back to rendering the text followed by the link in parens. This is fine,
but when we pass the full URL as the text, it's unnecessarily noisy, e.g.:
https://example.com/foo/bar (https://example.com/foo/bar)

Passing `fallback: false` toggles the fallback behaviour to just render the text as-is.

This wasn't clear previously from the docs:
https://github.com/sindresorhus/terminal-link/pull/26.