fix: tell users to re-run `netlify dev` after interrupting dev to init project
When a user runs `netlify dev` on an unlinked repository and they're
using an extension that requires a linked repository (e.g. Neon), we
interrupt them to initialize a project. Because of the way we cache
information about the current site, we're unable to effectively
invalidate the site state before starting the dev process, and so we
exit and expect the user to run `netlify dev` again to proceed.

We currently don't explicitly tell the user we're about to exit and
that they need to re-run `netlify dev`, though, which is jarring and
leaves the user in a bit of a lurch.

This changeset adds an informational message after `netlify dev` exits,
telling them to re-run the dev command. Ideally we would not require
this step, but we need to do some work on how we invalidate cached site
data before we can do that.